+++ packages/i18n/src/logs/logger.ts
@@ -1,0 +1,17 @@
+export default {
+  warn(message: string) {
+    console.warn(message);
+  },
+
+  error(message: string) {
+    console.error(message);
+  },
+
+  info(message: string) {
+    console.info(message);
+  },
+
+  debug(message: string) {
+    console.debug(message);
+  },
+};